% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotnet.R
\name{plotnet}
\alias{plotnet}
\title{Plot the wavelet correlation network.}
\usage{
plotnet(mat)
}
\arguments{
\item{mat}{a association strength matrix.}
}
\value{

}
\description{
Plot the wavelet correlation network.
}
\examples{
plotnet(mat = cor(Fauna.data[,-12]))
}
