\name{sawcn-package}
\alias{sawcn-package}
\alias{sawcn}
\docType{package}
\title{
\packageTitle{sawcn}
}
\description{
\packageDescription{sawcn}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sawcn}
\packageIndices{sawcn}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{sawcn}

Maintainer: \packageMaintainer{sawcn}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
