% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specwt.R
\name{specwt}
\alias{specwt}
\title{Calculate the wavelet transform coefficient.}
\usage{
specwt(series, mat)
}
\arguments{
\item{series}{like plot information or time series.}

\item{mat}{samples corresponding to plot or time.}
}
\value{

}
\description{
Calculate the wavelet transform coefficient.
}
\examples{
specwt(series = 1:100, mat = matrix(runif(100*6), nrow=100))
}
