% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w.gcor.R
\name{w.gcor}
\alias{w.gcor}
\title{Calculate the global wavelet correlation.}
\usage{
w.gcor(data, w, scales, reps)
}
\arguments{
\item{data}{the dataset of species and locations.}

\item{w}{the result of species dataset calculate by 'specwt'.}

\item{scales}{the scale selected to analysis.}

\item{reps}{times the bootstrap were performed.}
}
\value{

}
\description{
Calculate the global wavelet correlation.
}
\examples{
w.gcor(data = Fauna.data[,-12], w = specwt(1:119, Fauna.data[,-12]), scales = 57, reps = 500)
}
